Relocate data classes to model package
