+++ app/src/main/java/sa/edu/tuwaiq/jaheztask01/data/repositoryimp/FirebaseRepositoryImp.kt
@@ -3,7 +3,6 @@ package sa.edu.tuwaiq.jaheztask01.data.repositoryimp
 import android.util.Log
 import com.google.firebase.auth.AuthResult
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.auth.UserProfileChangeRequest
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
