+++ app/src/main/java/sa/edu/tuwaiq/jaheztask01/domain/model/ProfileState.kt
@@ -1,4 +1,4 @@
-package sa.edu.tuwaiq.jaheztask01.presentation.profile
+package sa.edu.tuwaiq.jaheztask01.domain.model
 
 import sa.edu.tuwaiq.jaheztask01.domain.model.RestaurantItem
 import sa.edu.tuwaiq.jaheztask01.domain.model.User
